passed all check but python

+++ geometry/util/math/mathutil.h
@@ -106,8 +106,8 @@ class MathUtil {
     // Discriminants below kTolerance in absolute value are considered zero
     // because changing the final bit of one of the inputs can change the
     // sign of the discriminant.
-    const double kTolerance = epsilon * max(fabs(2 * b * b), fabs(4 * a * c));
-    return (fabs(discriminant) <= kTolerance);
+    const double kTolerance = epsilon * max(fabsl(2 * b * b), fabsl(4 * a * c));
+    return (fabsl(discriminant) <= kTolerance);
   }
 
   // Returns in *r1 and *r2 the roots of a "normal" quadratic equation
@@ -393,7 +393,7 @@ class MathUtil {
 
   // the sine cardinal function
   static double Sinc(double x) {
-    if (fabs(x) < 1E-8) return 1.0;
+    if (fabsl(x) < 1E-8) return 1.0;
     return sin(x) / x;
   }
